Bugfix: Let sim note not interfere with title line
The similarity notice in the GistCard affected the position of the feed
and author title line, even if it wasn't present (i.e. in the GistList).

By setting the `gridTemplateColumns` correctly (it had a hardcoded
value) from a previous experiment and adding a `columnGap` only if the
similarity is present, this is fixed.