Add cleanup_bot to remove old gists
If a gist was removed from a feed, for example if the title and link
were changed, we still stored it in the database and displayed it in the
web UI. This caused problems as the updated entry is inserted again,
causing somewhat duplicate entries.

To counter this, this commit adds another service: cleanup_bot.
The cleanup_bot iterates over all gists every 5 minutes, checks whether
the link is still valid. This is only the case, if the original link
directly shows the page whithout redirection. If not, the gist is set to
"disabled". Then, it won't be displayed in the UI.

If this state is only temporary, e.g. if the news site decides to reply
with status code 200 again, the cleanup_bot enables the gist again.

This commit also adds a new error message in case the GistInspector is
opened for a disabled gist. This is possible when going through the
Telegram messages which are not retroactively removed in case of a
disabled gist.

Breaking changes: This commit adds a new column to the gists table:
"disabled".