Bugfix: Make author implicitly optional
Some feeds don't provide an author / creator for their entries. This can
either be simply "None" when parsing, or an empty string. To make this
consistent, we add that if the author / creator field is not present, we
return an emtpy string by default. The UI can already handle this.