Bugfix: System timezone not correctly displayed
If the timezone provided by the system is not one of the defaults
defined in the `react-timezone-select` library, it could not display it
in the timezone selector as it was not one of the defined options.

By first parsing the system timezone and then extracting the value again
before handing it into the Select component, this bug is fixed.