Increase delays and fetch feeds first in cleanup
To reduce the load on the news sites, we increase the gist fetch cycle
from 1 to 5 minutes. We also increase the cleanup cycle from 5 to 10
minutes.

Furthermore, previously we parsed the RSS feed for every gist in the
cleanup script. Instead, for each cycle we now fetch each feed once and
reuse the information.

This is done because one news site has already banned my IP address,
probably for DDoS protection.

resolves #14