Filter similar gists by disabled flag and feed id
This adds a disabled flag and the feed id to the insertion into the
chromadb. Furthermore, the query for similar gists now filters for
this flag and the feed.

Breaking change: The chromadb needs to be recreated or the documents
need to be updated manually so that they contain both the flag and the
feed ids.

resolves #13