Enable cleanup_bot to ensure chromadb completeness
The last change added the disabled state and the feed_id to the metadata
in the chromadb. This change adds the ability to the cleanup_bot to
check whether these properties are present in the chromadb. If they are
not, the cleanup_bot assures the correct data to be present.

Without this addition, the previous change, the added fields in the
metadata, does not work with previously added data.