Bugfix: Fix name mismatch in disabledFeeds state
In a recent change, the name of the state property for the disabled
feeds was partly changed to `disabledFeedIds`. "Partly" was the problem
though, as this name change has not been executed through the whole
frontend code.

This commit reverts this changed and sets the name of the property in
all occasions to `disabledFeeds` again. This fixes the problem that the
`disabledFeeds` state was not stored between sessions.