Unescape HTML Character entities in title
HTML entities are sanitized by default by the feedparser library to
prevent injection vulnerabilities when rendering the result in a
webpage. This for example causes all ampersands to be escaped as
`&amp;`. This looks a bit odd when rendered in the frontend.

As we always sanitize the input when using JSX templated objects, we can
safely unescape all relevant sanitized properties of the feeds and
entries. This commit does this for the titles of both the feeds and
titles.

resolves #6