Add news summaries over multiple days
This adds the new "recap" feature. Recaps of both the last 24 hours and
the last 7 days are generated on a daily basis. These recaps are stored
in the database and can be accessed from the frontend.

The recaps are constructed of multiple categories and each category
references up to 5 related gists.

resolves #23