Bugfix: Default response None if no recap in DB
Previously, the handler responded with todays date a year ago, if no
recap was found in the db when checking for the newest created date.
This broke the new logic to decide whether a recap is necessary.

This commit fixes this issue by responding with None of no recap was
found.