Add multi language support
This commit introduces multi-language support to the application. There
are two languages supported: German and English. The user can now select
their preferred language mode from the settings menu in the frontend or
stay with "Original" to keep the language of the original gist.

This introduces breaking changes to the database schema, updating
requires a full database reset and re-initialization.